Add optional help-type argument to help command line option:
The '-h' and '--help' command line options now support an optional "help type" argument identifying the type of help desired, such as version and/or build options information (i.e. "With:" and "Without:" list). The default is still just syntax, as it was before.

This allows one to determine not only just the version and nothing else, but also just information pertaining to how Hercules was built (i.e. what optional features it was built "With" and "Without").