Fix incorrect ORB Incorrect Length Indication Suppression handling:
Incorrect Length Suppression mode only occurs when both the Incorrect Length Suppression Mode bit and CCW Format Control bits are both on in the ORB, and then is supposed to apply to Immediate operations only, not non-Immediate operations. Due to a bug in channel.c it was being erroneously applied to non-Immediate operations as well. This patch fixes the bug.

Resolves Hercules-390 Hyperion Issues https://github.com/hercules-390/hyperion/issues/46 and https://github.com/hercules-390/hyperion/issues/161.