Update opcode table for facility instructions:
When a facility is disabled we update the opcode table for the instructions to route directly to program check operation exception. When the facility is enabled we restore the opcode table to its original value (route directly to the instruction function). This relieves each instruction function from having to do a separate FACILITY_CHECK each time it is executed allowing the instruction to execute more quickly.