Workaround for runtest crash only occurring on Linux:
For some as-yet unknown reason something the "stidp.tst" script does seems to cause whichever .tst script which follow it to crash (segfault) on Linux. Introducing a short delay seems to workaround the problem but why it does is unknown. This only occurs on Linux. On Windows everything always works just fine.