Fix architecture build order dependency bugs:
Jan's bug in cpu.c and sie.c are BOTH *almost* 16.5 years old (10 months short of a true day one bug). My identical bug in config.c and facility.c are both only a few months old.  We've been lucky so far due to the code in featchk.h (lines 623-649) that always builds 370 first, 390 second and 900 last.  But technically there *shouldn't* be any reason why they couldn't be built in *any* order.  Hercules should *not* be dependent on the build order of the various architectures it supports.  The code this patch fixes however, *was* coded to be dependent on the build order.  Hence the need to fix it.