Fix bug in hRexx "IsEnabled" function causing crash:
When hRexx's internal package name is initialized to "auto" (as a result of the HREXX_PACKAGE environment variable being defined to "auto"), the "IsEnabled" function erroneously returns 'true' when called for package name "auto", leading to a crash when the unstarted/uninitalized (non-existent!) interpreter function is eventually called to process the script.

This commit adds code to validate the package name passed to the "IsEnabled" function so that it now correctly returns 'false' when called for package name "auto", triggering the proper automatic enable/start logic to be called as a result, thereby preventing the crash from occurring.

closes #128