CCKD64 Support! (part 2 of 2):
Part of ongoing CCKD64 effort. This is the final commit that contains the actual CCKD64 support coding changes. Compare it with the part 1 of 2 commit to see what changes were actually needed to provide support for CCKD64.

THIS SHOULD BE CONSIDERED AN *EXPERIMENTAL* "BETA" RELEASE until such time as others have had a chance to give it a good workout to try and shake out any bugs that might still remain.

Please report any bugs to GitHub Issue #20:

    "We need 64-bit CCKD support!"
    (https://github.com/SDL-Hercules-390/hyperion/issues/20).

Thanks! (and enjoy!)

   :)