Improve PTTRACE accuracy and usability:
Previously, a 'ptt' command would show "hinlines.h" as being the location where intlock was acquired instead of where the e.g. OBTAIN_INTLOCK macro was used. This fixes that so that it now shows where the e.g. OBTAIN_INTLOCK macros was actually used as being the location where intlock was acquired.