Refresh all ExtPkg libs in light of recent cpu architecture fix:
Due to a recently discovered CMake build script bug in the External Packages, the actual cpu architecture of the resulting static libs were all the same (i.e. the so-called 32-bit libs, named libcrypto32.a, libdecNumber32.a, libSoftFloat32.a and libtelnet32.a, were actually 64-bit, not 32-bit).

This commit is to simply update all of the extpkg libs with the latest version, which, having now all been built with the fixed CMake script, are now all for the correct cpu architecture. (Only the *nix '.a' static libs were affected, but I'm committing new versions of all of the Windows VS2017 '.lib' files too in order to pick up the new "HHC01417I Built with..." version strings.)