Re-architect version/build information string handling:
Build version/information strings (messages) ONCE, and save for later use or display when needed. This will be leveraged in an upcoming commit to ensure minidump crash dumps include and report COMPLETE and accurate version/build information.

All of this is to try and gather additional information in an effort to try and eventually track down a crash that sometimes occurs (RARELY, but *sometimes*) somewhere in the cckd garbage collector's relocate Level-2 (L2) table logic, as reported in GitHub Issue #187.