(Oops!) Fix bass-ackwards operand register use in DIAG F09:
Operand-1 of an instruction is the *target* of the instruction (i.e. where the instruction's *results* are placed, i.e. the operand that the instruction *modifies*), whereas operand-2 and operand-3, etc, are *input* to the instruction's execution containing values used to determine *how* the operand-1 results should be calculated!

My original coding had the two operand registers (r1/r3) reversed! (Oops!) Sorry about that.