Fix special case vmfplc2 utility LOAD bug:
When loading a Variable length Textual file (such as an EXEC), if the 2-byte (HWORD) record length and the actual record data itself is split across separate blocks (i.e. a tape block ends with the 2-byte (HWORD) record length and the next tape block starts with the actual record), the record length would be forgotten and the first two bytes of the next block would be erroneously used as the record length, resulting in a garbled file being loaded onto the host system. This commit is the fix.

[skip travis]