Fix RARE crash in CCKD/CCKD64 "cckd(64)_sf_stats" function:
Problem was caught via user supplied crash dump.  IF, for whatever reason (the cause is immaterial), the 'fstat()' function happens to fail, the "st" variable (struct stat) is left in an uninitialized state, leading to a "Integer Divide-by-zero" exception (crash) during the calculation of the free percent value (how much of the dasd file is free expressed as a percentage of the file's size) when "st_size" field (file size in bytes) happens to be zero (due to it not being initialized because of the 'fstat()' system call failure).

Interestingly, this bug has been in Hercules since DAY ONE, since the VERY FIRST COMMIT made by Jay Maynard back on February 23, 2001:

   Revision: b95ef23f81f6220bdd9bf65e64d07911b57d89d3
   Author: Jay Maynard <jmaynard>
   Date: 2/23/2001 12:49:19 PM
   Message: Initial repository setup