Workaround fix for poor z/VM z/Arch guest performance:
This is only a temporary workaround fix. The actual/proper fix involves a major overhaul of the cpu.c module itself and slightly different code in sie.c

Technically this resolves GitHub Issue 222, but we're going to keep it open for a while to make sure we didn't accidentally break something else (i.e. we want to wait for more user feedback before officially closing issue 222).