MAINLOCK Corrections and Improvements
The Interlocked Access Facility 1 & 2 instructions were corrected in the previous commit (1c54231) by protecting them using OBTAIN_/RELEASE_MAINLOCK.

In order to optimize MAINLOCK usage where possible, assists for cmpxchg1, cmpxchg4 and cmpxchg8 are now created in case C11 atomics are available.  Assists for cmpxchg16 in the case of x86_64 for Windows, Linux and MacOS are also created, as well as for the aarch64 architecture (AMRv8 such as used by the Raspberry Pi 4).

The first argument used by cmpxchg16 needs to be 16-byte aligned in certain cases.  The CDSG and CSST instructions have been corrected accordingly.  Also the STPQ and LPQ instructions now use cmpxchg16.

An _UNCONDITIONAL version of MAINLOCK is created identical to the existing one.  MAINLOCK itself can be nullified in case all cmpxchg1/4/8/16 assists and H_ATOMIC_OP are available.  Only the PLO instruction needs the _UNCONDITIONAL version of MAINLOCK.