Reset NUMCPU back to default 1 after CDSG test!
Failure to do so was causing the next following test (in this case the 'cipher.tst' test) to always fail since it was written to always expect the default startup number of CPUs (which is 1).

If any of your runtest .tst tests need to specify a NUMCPU value > 1, then PLEASE e sure to reset it back to 1 again after your test completes!