Restore erroneously removed printer immediate CCW opcodes tables:
Commit bff81c3896ee08c20d61b0690315803c5133e3f9 ("Fix 1403 printer Load UCS bug") removed the CCW "immed" tables because, according to the change log:

   2. Some CCWs were being erroneously processed by the
      channel as "immediate" CCWs instead of being passed
      to (and processed by) the printer driver itself.

but: 1) it doesn't say which CCWs they were, and 2) whether or not a CCW is immediate or not has NOTHING to do with whether or not it is passed to the driver! (*)

BOTTOM LINE: The immediate CCW opcode table should NOT have been removed and causes unexpected Channel Program Checks in certain conditions. This commit corrects the error.

(*) *ALL* CCWs -- whether they are immediate or not -- must *always* passed to the driver! Duh! I don't know what in the world I was thinking when I originally made that [completely bogus!] change! Sorry!

[skip travis]