Restore watchdog support:
The old watchdog thread detected hung CPUs and sent a sigabend signal that was caught and threw a machine check for the malfunctioning CPU in order to try and allow the guest to continue running. This was removed months ago (April 2019) by commit b76163a6646abef539e6800df3fbdb6adaed9a9d.

The new watchdog thread not only detects and reports hung CPUs as before, but now also FIRST detects and reports deadlocked threads too, and does NOT attempt to allow the guest to continue (does NOT cause a Machine Check). Instead, it creates a crash dump for offline post mortem analysis.