Disable DEBUG_FREESPACE build option:
The DEBUG_FREESPACE #define was meant only for debugging purposes during code development only (such as during the development of CCKD64 support), and was NOT meant to remain enabled for normal release/production builds.

This resolves a user-reported problem where the log file was flooded with literally millions of HHC00398I (and HHC00399I) internal cckd trace messages because of cckd logic having decided to dynamically consolidate/remove a volume's free space due to it becoming "too fragmented" (which under normal circumstances does not happen very often, if at all, which is why it took so long for this bug to be caught).