Fix incomplete network device group crash:
If a network device (CTCI, LCS, QETH, PTP) is incorrectly defined with the wrong number of devices for the group such that one or more of the groups was left "incomplete" (e.g. "0E20.3 CTCI ..."), Hercules would immediately crash in the "device_reset" function during IPL whenever it tried to reset the non-existent device. This is the fix.

Thank you to Martien van Balgooi for reporting this problem!