Instruction decoders cleanup part 8: IBM formats:
Update instruction comments with actual IBM formats, as well as in the new GENx370x390x900 macro argument (opcode.c column) too.

Some instruction functions moved the *follow* the OPTION_OPTINST optimized functions instead of preceding them (for easier parsing(*)).

That's pretty much it for right now.

More to come in future commits as this project progresses.

-----------------
(*) I've written a custom tool to help me with this project and doing so makes it easier for my program to deal with (parse) Hercules code.