Clean up overall ARCH_DEP structure of hscmisc.c module:
The overall layout/structure (flow) of the hscmisc.c source module was all wrong. There were non-ARCH_DEP functions in the ARCH_DEP section of the module and vice-versa. All ARCH_DEP source modules (i.e. those that are compiled (#included) multiple times under control of the _GEN_ARCH preprocessor macro, once for each supported build architecture) SHOULD follow the same basic layout as documented in the "_archdep_templ.c" example/sample dummy source module. That's why I created it.

Therefore this change (commit) should be considered mostly "cosmetic" in nature/spirit/intent, even though technically it obviously is not. That is to say, no actual functionality was changed. It should still function and behave as before. This was just a "code cleanup" effort, a reorganizing of the code, and nothing else.

[skip travis]