(guaranteed 100% cosmetic!) [skip travis]
Hercules uses makefiles to build with and thus does not use the Visual Studio project's preprocessor defines. It only uses the values specifically /D defined in its makefile or #defined in the code itself. This change is only so the Visual Studio editor's "Intellisense" feature doesn't visually "disable" (deactivate/gray out) active code blocks because it thinks it's inactive code (unused/disabled). Thus this commit is 100% cosmetic, I guarantee it!