Various TXF fixes:
1. Fix stupid Fish TXF DAT bug (arn == USE_REAL_ADDR) (Thank you, Peter!) (And apologies to Bob Wood! YOU had it RIGHT and I ended up breaking it!(*) Sorry!) (Note: actual fix was already previously committed accidentally by commit fa645447fc11e12ab3ff27ad73833ed8e506dfc2)

2. Disallow nested constrained transactions!

3. Allow constrained transaction to be nested within an unconstrained transaction, and ensure correct Program Interrupt Filtering flags (PIFC) are used when that occurs.

4. Fix careless Fish typo in TDB logic (Thank you, Jürgen!)

[skip travis]

(Skipping travis only because this is just the first commit of several commits. The last commit will allow travis to run.)

---------------------
(*) At the time I was completely misunderstanding the purpose of the 'USE_REAL_ADDR' flag and my disabling of the arn test was only meant to be temporary until I could complete additional research, which I never got around to doing, thereby accidentally committing something that was never intended to be committed!