Stop writing after ETX / ETB in non transparent mode
The IBM 2703 Transmission Control Component Description document specifies on
page 58 that a channel write shall stop if an ETX or ETB control character
is detected in the data stream.

It is also mentioned on page 48 in the same document that bytes following the ETX or ETB shall
not be transferred to the line if they are received from the channel.

This fix detects the presence of these characters and breaks out of the loop
if these occur in the data stream.

___
The bug fixed by this commit was discussed extensively on hercules-390 groups.io
group in the "Yet more problems with BSC in commadpt.c - ACK response lost" thread:
https://hercules-390.groups.io/g/group/topic/yet_more_problems_with_bsc_in/75630306

This fix was originally offered by Mattis as a Hercules 3.13 spinhawk pull request:
https://github.com/rbowler/spinhawk/pull/93/commits/2d93c3db0b93aacc9ba62845ed82b328550d7cb9