Don't reset instruction fetch constraint if nested constrained:
Yes, constrained transactions cannot be nested, but they CAN be called as part of an unconstrained transaction, and when that happens, the "constrained" trasnaction is not constrained. It just becomes another nesting level of the already unconstrained transaction, which means the instruction fetch constraint should NOT be reset.

[skip travis]

(Skipping travis because this is just one of several commits in a long series of commits. The last commit in the series will allow travis to run.)