Walk region/segment/page tables in HW Load function:
Function 'hwl_loadfile' in scescsi.c fixed to load requested file by properly walking the Region and/or Segment and Page tables. Closes GitHub Issue #185 and Pull Request #337.

Note: overall fix was developed by Erik, but the logic is actually Fish's, so blame Fish if any problems are discovered! I've committed the fix under his name (with his permission) so that he gets overal credit for the fix since it was his hard work that determined what the problem was and how it should be fixed. Thank you, Erik!