More proper(?) fix for GitHub Issue #340
This is a more proper fix for GitHub Issue #340, which itself identified a problem with commit 7a52a4bbd7e574f583017462454e1c519c504623, so this is a more proper fix for commit 7a52a4bbd7e574f583017462454e1c519c504623 as well.

It enhances the OSTAILOR command in such a way as to make the test for z/VM in cpu.c a bit more reliable so as to resolve GitHub Issue #340.

I don't particularly like relying on an OSTAILOR setting to determine whether the guest is running z/VM or not, but we don't have much choice as there is really no other reliable way to do so.