Refresh all default ExtPkg libs and configure files:
This should be a cosmetic change for those who build the External Packages themselves.

For those who don't however (i.e. for those who build and link Hercules with the default ExtPkg lib files that come pre-delivered with the Hercules distribution, this should update all ExtPkgs to their latest and greatest version.