Make the call to recv on the telnet connection non-blocking.
The device lock is held when this call is made and if it blocks
things go down hill in a hurry.  According to the Linux select(2)
man page a call to recv may block even a previous call to select
says it shouldn't.  This has also been seen on MacOSX.