Make the call to recv on the telnet connection non-blocking. (#379)
* Make the call to recv on the telnet connection non-blocking.

The device lock is held when this call is made and if it blocks
things go down hill in a hurry.  According to the Linux select(2)
man page a call to recv may block even a previous call to select
says it shouldn't.  This has also been seen on MacOSX.

* Make the 1052 socket non-blocking too, it can have the same issue.

* Ignore EAGAIN errors on 1052 reads.  Just return zero and try again later.