Corrections to autoconf for BSD 12 shared libraries
The autoconf/automake, etc. included with Hercules are 15 to 20 years out of date and were never set up to handle newer FreeBSD properly.  As a result shared libraries and linking are not handled.  This change addresses that.

Also FORTRAN detection in configure has been known to goof up shared libraries as well.  So we disable its detection.