Fix build errors in gcc8 on MacOS Mojave, part 2 of 3
Include machdep.o in libhercu to avoid undefined symbols
such as _fetch_dw_noswap, _fetch_fw_noswap, etc.

This may not be the best fix for this, but it seems to work.  It
makes more sense to make libhercu depend on libherc which already
includes machdep, but this is a circular dependency.  Instead include
it in both libraries.