Fix build errors (#412)
* Fix build errors in gcc8 on MacOS Mojave, part 1 of 3

Fix compile errors in skey.c similar to

../../hercules/skey.c:124:22: warning: inline function 'bypass_skey_update' declared but never defined
   extern inline bool bypass_skey_update( REGS* regs, BYTE m3, BYTE oldkey, BYTE r1key );

* Fix build errors in gcc8 on MacOS Mojave, part 2 of 3

Include machdep.o in libhercu to avoid undefined symbols
such as _fetch_dw_noswap, _fetch_fw_noswap, etc.

This may not be the best fix for this, but it seems to work.  It
makes more sense to make libhercu depend on libherc which already
includes machdep, but this is a circular dependency.  Instead include
it in both libraries.

* Fix build errors in gcc8 on MacOS Mojave, part 3 of 3

Add extern inline declarations for _get_storekey_ptr,
_get_dev_storekey_ptr, etc. to skey.c so they aren't  undefined
suring linking.