Fix another stupid-compiler warning:
Any by "stupid" I mean the compiler(s), not necessarily the warning. The warning is not stupid (except in the case where it's issued when it shouldn't be), but the compiler is certainly stupid for issuing the bogus warning instead of taking the time to determine conclusively whether it's actually warranted or not.