SYSGPORT tweak: "SYSGPORT NO" is now the default:
This *should* provide original behavior when no SYSGPORT statement is given, while providing the new behavior when a SYSGPORT statement *is* given [with a valid port number(*)].

This *should* now PROPERLY close GitHub Issue #505.

--------------
(*) Same as current CNSLPORT behavior, if an invalid port number is specified, a default value is used (which for SYSGPORT is 3278).