Fix tape locateblk_virtual locking issue:
See GitHub Issue #518. If Locate Block takes "too long"(*), due to current device lock handling in the locateblk_virtual function, it's possible for a processor to hang should it execute certain I/O instructions (due to inability to acquire the device lock). This commit improves device lock handling to prevent such occurrences and closes #518.

(*) Locate Block (locateblk_virtual function) *should* be very fast and under normal circumstances should *not* take "too long". But in this particular case it was taking 35(!) seconds. It is unknown why it was taking so long for this user (it's being looked into!) but if it ever does happen, the current fix being committed *should* prevent any processor hang from occurring.