Improve PFPO instruction (partial almost fix):
NOTE: it's still not right (i.e. it still contains known bugs), but this is the best we can do for now. The condition code is at least now being set (probably not correctly in all situations though) as well as the Return Code value in GR1 too (same thing: probably not correctly in all situations), but it's an improvement over what we had before which wasn't setting either one!

It's going to take a non-insignificant amount of time and effort to get it completely fixed, but for now this should HOPEFULLY suffice.