NEW FEATURE: "Trace to File" support:
New 'tf' command to direct all instruction and I/O tracing to a disk file instead of immediately formatting and displaying on the console (which slows down Hercules execution to a CRAWL). Only the raw data is written to the trace file and a new 'tfprint' utility is provided to produce the actual trace report (identical to what would normally be displayed on the Hercules console).

Closes issue #33