Clear up misconceptions regarding "SDL" Hercules 4.x Hyperion:
Make appropriate updates to documentation, images and text to make very clear that this is NOT a "SoftDevLabs" version/product of Hercules! The "SDL" in its name is just a name, and is used to distinguish it from the other versions of Hercules that are out there. Nothing more. The existence of "SDL" in its name does NOT mean to imply Hercules is in any way associate with Software Development Laboratories. (I have also updated my own web site as well to try and make that clear.)

HERCULES IS NOT MINE! Hercules belongs to the Hercules project. Hercules belongs to the community. I am simply a member of the Hercules development team, that's all. A single contributor. Nothing more.