Fix fpr regs display so that it *always* works:
Previously, fpr regs would be incorrectly displayed depending on whether the CR0_AFP was set and/or whether the HERC_370_EXTENSION facility was enabled. Now it works correctly *regardless* of either setting.

The problem was it was directly accessing the 'fpr' array entries manually, instead of using the provided 'FPR2I' macro like it should be doing. This has been fixed.

I also tweaked the actual display format a tad too: I didn't like the confusing order it was using. Now it more sensibly displays 0..7 down the left column, and 8..15 down the right column.