(build member updates missed by commit c67b47ff6e51cf014c3f974ea32fa0efbc0636aa)
The "Implement Message Security Assists 5 and 7" commit missed adding the test files to the Linux 'tests/Makefile.am' build member as well as to the various Visual Studio project files. This commit makes those necessary changes.

(Don't worry about it Jürgen! We're all part of a team, and we all watch each's backs! Right? It'd no big deal. I'm happy to take care of it for you!)