minor cosmetic message fix:
I noticed the "Checking for" message for -minline-stringops-dynamically was being issued in the middle of the "Checking for or what optimization flags to use" message instead of on it's own separate line BEFORE the optimization flags message. This fixes that so that they're correctly issued separately on their own separate line.