100% COSMETIC update to Visual Studio 2008 preprocessor defines:
Hercules uses makefiles to build with, NOT Visual Studio project files. The only preprocessor defines that Hercules uses are those specifically defined in its makefile or passed on the command line (or #defined in its source files of course).

This change is ONLY to prevent the Visual Studio 2008 editor from disabling (deactivating) regions of code delimited by #if/#endif etc, thereby preventing its "Intellisense" feature from behaving correctly within that region (because it thinks that section of code isn't active due to the #if guarding it).

This change simply makes it easier to work with Hercules source coding using the Visual Studio 2008 editor, nothing more. It does *NOT* impact the Hercules build process in ANY way whatsoever!!